+++ tgui/packages/tgui/components/Tooltip.tsx
@@ -53,6 +53,10 @@ export class Tooltip extends Component<TooltipProps, TooltipState> {
       <Popper
         options={{
           placement: this.props.position || "auto",
+          modifiers: [{
+            name: "eventListeners",
+            enabled: false,
+          }],
         }}
         popperContent={
           <div
