+++ tgui/packages/tgui/interfaces/AlertModal.js
@@ -4,7 +4,7 @@
  * @license MIT
  */
 
-import { clamp01 } from 'common/math';
+import { Loader } from "./common/Loader";
 import { useBackend } from '../backend';
 import { Component, createRef } from 'inferno';
 import { Box, Flex, Section } from '../components';
@@ -135,15 +135,3 @@ export class AlertModal extends Component {
   }
 
 }
-
-export const Loader = props => {
-  const { value } = props;
-
-  return (
-    <div className="AlertModal__Loader">
-      <Box
-        className="AlertModal__LoaderProgress"
-        style={{ width: clamp01(value) * 100 + '%' }} />
-    </div>
-  );
-};
